:truck: Move some components and functions
