+++ .env.sample
@@ -3,3 +3,5 @@ PUSHER_APP_ID=<your_pusher_app_id>
 PUSHER_KEY=<your_pusher_key>
 PUSHER_SECRET=<your_pusher_secret>
 PUSHER_CLUSTER=<your_pusher_cluster>
+NEXT_PUBLIC_PUSHER_KEY=<your_pusher_app_id>
+NEXT_PUBLIC_PUSHER_CLUSTER=<your_pusher_cluster>
