+++ app/room/[randomName]/page.tsx
@@ -2,8 +2,6 @@
 
 import {AwardIcon} from "lucide-react";
 import {Card, CardContent, CardDescription, CardHeader, CardTitle} from "@/components/ui/card";
-import React from "react";
-
 
 export default function CodeReviewQuiz({params}: Readonly<{ params: { randomName: string } }>) {
     const randomName = params.randomName;
