+++ components/RealTimeRoom.tsx
@@ -5,7 +5,7 @@ import {Card, CardContent} from "@/components/ui/card";
 import {Code2, Shuffle} from "lucide-react";
 import {AnimatePresence, motion} from "framer-motion";
 import {Button} from "@/components/ui/button";
-import {openAnonymousRandomRepositoryServerActionV2} from "@/src/backend/server";
+import {openAnonymousRandomRepositoryServerActionV2} from "@/lib/backend/server";
 
 const RealTimeRoom = () => {
     const [repositories, setRepositories] = useState<string[]>([]);
