+++ lib/backend/realTimeManager.ts
